Add Module To Remove Task At Index
