+++ src/index.html
@@ -16,7 +16,7 @@
 
       <div class="todo-title">
         <h2>Today's To Do</h2>
-        <button type="submit"><i class="fas fa-sync fa-xs"></i></button>
+        <button onClick="window.location.reload()"><i class="fas fa-sync fa-xs"></i></button>
       </div>
  
       <!-- Add ToDo form -->
