+++ src/index.js
@@ -1,38 +1,10 @@
 import './style.css';
-import completeTask from './complete.js';
-import { setLocalStorage, addToLocalStorage } from './setLocalStorage.js';
-import getTasksFromLocalStorage from './getTasks.js';
+import showTasks from './showTasks.js';
+import { addToLocalStorage } from './setLocalStorage.js';
+import deleteTodo from './removeOneTask.js';
 
-// Dynamically display tasks
+// Global Variables
 const listContainer = document.getElementById('todo-lists');
-
-const showTasks = () => {
-  listContainer.innerHTML = '';
-  const tasks = getTasksFromLocalStorage();
-  for (let i = 0; i < tasks.length; i += 1) {
-    const list = ` <li class="task" id="${tasks[i].index}" draggable="true">
-    <div>
-      <input type="checkbox" class="box" id="list-box" name="list-box">
-      <label class="form-label">${tasks[i].description}</label>
-    </div>
-    <button class="ellipsis"><i class="fas fa-ellipsis-v fa-xs"></i></button>
-    <button class="delete"><i class='fas fa-trash-alt'></i></button>
-  </li>`;
-
-    listContainer.innerHTML += list;
-  }
-
-  // Add Line Through When Checkbox Is Checked
-  const box = document.querySelectorAll('.box');
-  for (let j = 0; j < box.length; j += 1) {
-    box[j].checked = tasks[j].completed;
-    box[j].addEventListener('change', (event) => {
-      completeTask(event.target, tasks[j]);
-      setLocalStorage(tasks);
-    });
-  }
-};
-
 const input = document.querySelector('.text');
 
 function clearInput() {
@@ -62,31 +34,12 @@ const addTodoTask = (e) => {
 const form = document.getElementById('form');
 form.addEventListener('submit', addTodoTask);
 
-// Make Label Editable On Double Click
-const label = document.querySelectorAll('label');
-label.forEach((item) => {
-  item.addEventListener('dblclick', () => {
-    item.setAttribute('contenteditable', 'true');
-  });
-
-  // Show Delete Icon On Focus
-  item.addEventListener('focus', () => {
-    const last = item.parentElement.parentElement.lastElementChild;
-    last.style.display = 'block';
-    item.parentElement.parentElement.style.backgroundColor = '#dadadc';
-
-    const ellipsis = item.parentElement.parentElement.lastElementChild.previousElementSibling;
-    ellipsis.style.display = 'none';
-  });
-
-  // Hide Delete Icon On Blur
-  item.addEventListener('blur', () => {
-    const last = item.parentElement.parentElement.lastElementChild;
-    last.style.display = 'none';
-    item.parentElement.parentElement.style.backgroundColor = '#fff';
-    const ellipsis = item.parentElement.parentElement.lastElementChild.previousElementSibling;
-    ellipsis.style.display = 'block';
-  });
+listContainer.addEventListener('click', (event) => {
+  if (event.target.classList.contains('delete-task')) {
+    const listKey = event.target.parentElement.parentElement.dataset.key;
+    deleteTodo(listKey);
+    showTasks();
+  }
 });
 
 showTasks();
